Set timed_out attrib on reports for logging customization
In case the test timed out, set `report.timed_out = True`. If the test
ran without a timeout, set `report.timed_out = False`. In any other case
`report.timeout_out` should be set to `None` or be undefined.
That makes it easier to customize the output in case of timeouts.

This commit also adds a small section to the README including a small
example.